+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `GoWithID` starts a goroutine with a new request tracking ID.
+
+### Changed
+- `Go` no longer issues new ID automatically.  Use `GoWithID` instead.
 
 ## [1.2.0] - 2016-08-31
 ### Added